Buoyancy Controller added to Box2D (AndEngine/libgdx)
For details see Yannick Loriot's https://github.com/YannickL/Box2D-Examples

b2ChainShape needs a ComputeSubmergedArea to be defined (this didn't exist in version 2.2.0 which is used in YannickL's example).

JNI wrappers and bindings still need to be created for b2Controller, b2BuoyancyController and if required, ComputeSubmergedArea for each of the Shapes.